refactor: Cấu trúc lại thư mục chuẩn bị làm cho phần product for customer/dealer
