+++ app/Http/Controllers/Customer/ProductController.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace App\Http\Controllers\Customer;
+
+use Illuminate\Http\Request;
+
+class ProductController
+{
+    //
+}
