+++ app/Http/Controllers/System/DealerRequestController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\System;
 
 use Exception;
 use App\Traits\ApiResponse;
