+++ app/Http/Resources/Customer/ProductResource.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace App\Http\Resources\Customer;
+
+use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\JsonResource;
+
+class ProductResource extends JsonResource
+{
+    /**
+     * Transform the resource into an array.
+     *
+     * @return array<string, mixed>
+     */
+    public function toArray(Request $request): array
+    {
+        return parent::toArray($request);
+    }
+}
