+++ routes/api/v1/auth.php
@@ -1,7 +1,7 @@
 <?php 
 
 use Illuminate\Support\Facades\Route;
-use App\Http\Controllers\AuthController;
+use App\Http\Controllers\System\AuthController;
 
 
 Route::prefix('auth')->group(function(){
